\name{fgkm.sample}
\docType{data}
\alias{fgkm.sample}
\title{Sample dataset to illustrate the fgkm algorithm.}
\description{

  A sample dataset of 50 variables and 600 observations.

}
\usage{fgkm.sample}
\format{

  The \code{fgkm.sample} dataset is a data frame with 600 observations
  and 50 variables.

}
\seealso{

  \code{\link{fgkm}}.

}
\keyword{datasets}
