\name{twkm.sample}
\docType{data}
\alias{twkm.sample}
\title{Sample dataset to test the twkm algorithm.}
\description{

  A sample dataset of 410 variables and 2000 observations. The last variable is the class variable, and should be removed before clustering.
  The grouping of the variables are "0-75;76-291;292-355;356-402;403-408".
}
\usage{twkm.sample}
\format{

  The \code{twkm.sample} dataset is a data frame with 2000 observations
  and 50 variables.

}
\seealso{

  \code{\link{twkm}}.

}
\keyword{datasets}
